+++ src/Wt/WServer.C
@@ -317,7 +317,7 @@ void WServer::restart(const std::string &applicationPath,
 #ifndef WT_WIN32
   std::unique_ptr<char*[]> argv(new char*[args.size() + 1]);
   argv[0] = const_cast<char*>(applicationPath.c_str());
-  for (int i = 0; i < args.size(); ++i) {
+  for (int i = 0; i < static_cast<int>(args.size()); ++i) {
     argv[i+1] = const_cast<char*>(args[i].c_str());
   }
